requirements.txt
Pillow==8.4.0
reportlab==3.6.7
opencv-python==4.5.5.62